+++ ci/jenkins/step/shardsDevNightlyTest.groovy
@@ -13,6 +13,7 @@ timeout(time: 180, unit: 'MINUTES') {
                 def helmResult = sh script: "helm status -n milvus ${env.SHARDS_HELM_RELEASE_NAME}", returnStatus: true
                 if (!helmResult) {
                     sh "helm uninstall -n milvus ${env.SHARDS_HELM_RELEASE_NAME}"
+                    sleep 30
                 }
                 throw exc
             }