+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#2073 Fix CheckDBConfigBackendUrl error message
 -   \#2076 CheckMetricConfigAddress error message
 -   \#1796 Too much low performance of building index on ubuntu-mysql-version
+-   \#2141 Fix server start failed if wal directory exist
 
 ## Feature
 -   \#1751 Add api SearchByID
@@ -20,10 +21,12 @@ Please mark all change in change log and use the issue from GitHub
 -   \#2054 Check if CPU instruction sets are illegal
 -   \#2059 Add lock file avoid multiple instances modifying data at the same time
 -   \#2064 Warn when use SQLite as metadata management
+-   \#2111 Check GPU environment before start server
 
 ## Improvement
 -   \#221 Refactor LOG macro
 -   \#2039 Support Milvus run on SSE CPUs
+-   \#2149 Merge server_cpu_config.template and server_gpu_config.template
 
 ## Task